+++ pkg/namespaced/clients/projects/approvalrules_test.go
@@ -124,7 +124,7 @@ func TestIsUserIDsUpToDate(t *testing.T) {
 		},
 		"EmptyUserIDsEmptyUsers": {
 			cr: &v1alpha1.ApprovalRuleParameters{
-				UserIDs: &[]int{},
+				UserIDs: &[]int64{},
 			},
 			in: &gitlab.ProjectApprovalRule{
 				Users: []*gitlab.BasicUser{},
@@ -133,7 +133,7 @@ func TestIsUserIDsUpToDate(t *testing.T) {
 		},
 		"MatchingUserIDs": {
 			cr: &v1alpha1.ApprovalRuleParameters{
-				UserIDs: &[]int{1, 2},
+				UserIDs: &[]int64{1, 2},
 			},
 			in: &gitlab.ProjectApprovalRule{
 				Users: []*gitlab.BasicUser{
@@ -145,7 +145,7 @@ func TestIsUserIDsUpToDate(t *testing.T) {
 		},
 		"DifferentUserIDs": {
 			cr: &v1alpha1.ApprovalRuleParameters{
-				UserIDs: &[]int{1, 2},
+				UserIDs: &[]int64{1, 2},
 			},
 			in: &gitlab.ProjectApprovalRule{
 				Users: []*gitlab.BasicUser{
@@ -157,7 +157,7 @@ func TestIsUserIDsUpToDate(t *testing.T) {
 		},
 		"DifferentLengths": {
 			cr: &v1alpha1.ApprovalRuleParameters{
-				UserIDs: &[]int{1},
+				UserIDs: &[]int64{1},
 			},
 			in: &gitlab.ProjectApprovalRule{
 				Users: []*gitlab.BasicUser{
