+++ pkg/namespaced/clients/runners/runner.go
@@ -85,7 +85,7 @@ func GenerateGroupRunnerObservation(runner *gitlab.RunnerDetails) groupsv1alpha1
 	groups := make([]groupsv1alpha1.RunnerGroup, 0, len(runner.Groups))
 	for _, group := range runner.Groups {
 		groups = append(groups, groupsv1alpha1.RunnerGroup{
-			ID:     int(group.ID),
+			ID:     int64(group.ID),
 			Name:   group.Name,
 			WebURL: group.WebURL,
 		})
@@ -112,7 +112,7 @@ func GenerateProjectRunnerObservation(runner *gitlab.RunnerDetails) projectsv1al
 	projects := make([]projectsv1alpha1.RunnerProject, 0, len(runner.Projects))
 	for _, project := range runner.Projects {
 		projects = append(projects, projectsv1alpha1.RunnerProject{
-			ID:                int(project.ID),
+			ID:                int64(project.ID),
 			Name:              project.Name,
 			NameWithNamespace: project.NameWithNamespace,
 			Path:              project.Path,
@@ -135,7 +135,7 @@ func generateCommonRunnerObservation(runner *gitlab.RunnerDetails) commonv1alpha
 		return commonv1alpha1.CommonRunnerObservation{}
 	}
 	runnerObservation := commonv1alpha1.CommonRunnerObservation{
-		ID:              int(runner.ID),
+		ID:              runner.ID,
 		Description:     runner.Description,
 		Paused:          runner.Paused,
 		Locked:          runner.Locked,
@@ -147,7 +147,7 @@ func generateCommonRunnerObservation(runner *gitlab.RunnerDetails) commonv1alpha
 		Status:          runner.Status,
 		RunUntagged:     runner.RunUntagged,
 		AccessLevel:     runner.AccessLevel,
-		MaximumTimeout:  int(runner.MaximumTimeout),
+		MaximumTimeout:  runner.MaximumTimeout,
 		IsShared:        runner.IsShared,
 	}
 
@@ -186,7 +186,7 @@ func IsRunnerUpToDate(spec *commonv1alpha1.CommonRunnerParameters, observed *git
 		return false
 	}
 	// Convert observed.MaximumTimeout from int64 to int for comparison
-	observedMaxTimeout := int(observed.MaximumTimeout)
+	observedMaxTimeout := int64(observed.MaximumTimeout)
 	// Use a compact list to keep cyclomatic complexity low
 	checks := []bool{
 		clients.IsComparableEqualToComparablePtr(spec.Description, observed.Description),
